local

+++ client-angular/src/services/auth.service.ts
@@ -1,3 +1,4 @@
+import { isPlatformBrowser } from '@angular/common';
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable, tap } from 'rxjs';
@@ -8,13 +9,12 @@ import { Observable, tap } from 'rxjs';
 export class AuthService {
   private baseUrl = 'https://server-dotnet.onrender.com/api/Auth';
 
-  constructor(private http: HttpClient) {}
-
+  constructor(private http: HttpClient, private platformId: Object) {}
   login(credentials: { name: string, password: string }): Observable<any> {
     return this.http.post<{ token: string }>(`${this.baseUrl}/login`, credentials)
       .pipe(
         tap(response => {
-          if (response.token) {
+          if (response.token && isPlatformBrowser(this.platformId)) {
             localStorage.setItem('token', response.token);
             const payload = this.decodeToken(response.token);
             // const role = payload?.role || 'User';
@@ -23,14 +23,15 @@ export class AuthService {
             payload?.role || 'User';
             console.log(role)
             console.log(response.token)
-            if (role !== 'Admin') {
+            if (role !== 'Admin' && isPlatformBrowser(this.platformId)) {
               // לא מנהל – נזרוק שגיאה ונמחק את הטוקן
               localStorage.clear();
               throw new Error('גישה מותרת למנהלים בלבד');
             }
-  
+  if( isPlatformBrowser(this.platformId)){
             // שמירת פרטים אם עבר את הבדיקה
             localStorage.setItem('role', role);
+  }
             // sessionStorage.setItem('name', payload?.unique_name || '');
           }
         })
